feat: Implement copying input and result to clipboard
Option+Command+C (Mac) or Ctrl+Shift+C to copy **result** to clipboard.
The point is to use a more convenient (for the fingers) key sequence
to copy the result, which users are more likely to be interested in than
the input value.  Hence, this is why we use Ctrl+Shift+C to copy the
result to clipboard on Linux/Windows, despite one being able to argue
that Ctrl+Alt+C is closer to Option+Command+C on Mac than Ctrl+Shift+C.